drivers/scsi:ufs:add hikey970 ufs driver
Modify ufs driver to support both hikey960 and hikey970.
