+++ arch/arm64/configs/hikey960_defconfig
@@ -219,7 +219,7 @@ CONFIG_HISI_HIKEY_USB=y
 CONFIG_BLK_DEV_SD=y
 CONFIG_SCSI_UFSHCD=y
 CONFIG_SCSI_UFSHCD_PLATFORM=y
-CONFIG_SCSI_UFS_HI3660=y
+CONFIG_SCSI_UFS_HISI=y
 CONFIG_ATA=y
 CONFIG_SATA_AHCI=y
 CONFIG_SATA_AHCI_PLATFORM=y
