+++ drivers/scsi/ufs/Kconfig
@@ -80,14 +80,6 @@ config SCSI_UFSHCD_PLATFORM
 
 	  If unsure, say N.
 
-config SCSI_UFS_HI3660
-	tristate "Hisilicon Hi3660 UFS controller platform driver"
-	depends on (ARCH_HISI || COMPILE_TEST) && SCSI_UFSHCD_PLATFORM
-	help
-	  This selects the Hisilicon HI3660 additions to UFSHCD platform driver.
-
-	  If unsure, say N.
-
 config SCSI_UFS_DWC_TC_PLATFORM
 	tristate "DesignWare platform support using a G210 Test Chip"
 	depends on SCSI_UFSHCD_PLATFORM
@@ -107,4 +99,12 @@ config SCSI_UFS_QCOM
 	  specific registers.
 
 	  Select this if you have UFS controller on QCOM chipset.
+	  If unsure, say N.
+
+config SCSI_UFS_HISI
+	tristate "Hisilicon Kirin UFS controller platform driver"
+	depends on (ARCH_HISI || COMPILE_TEST) && SCSI_UFSHCD_PLATFORM
+	help
+	  This selects the Hisilicon Kirin additions to UFSHCD platform driver.
+
 	  If unsure, say N.
