+++ drivers/scsi/ufs/ufs_quirks.h
@@ -144,6 +144,12 @@ struct ufs_dev_fix {
  */
 #define UFS_DEVICE_QUIRK_HOST_PA_SAVECONFIGTIME	(1 << 8)
 
+/*
+ * Some UFS devices require VS_DebugSaveConfigTime is 0x10,
+ * enabling this quirk ensure this.
+ */
+#define UFS_DEVICE_QUIRK_HOST_VS_DEBUGSAVECONFIGTIME	(1 << 9)
+
 struct ufs_hba;
 void ufs_advertise_fixup_device(struct ufs_hba *hba);
